Refactor progress button
